piid-imgzipredaction/ImageRedaction/requirements.txt
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
enum34==1.1.6
Flask==1.0.2
futures==3.2.0
google-api-core==1.5.2
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-dlp==0.9.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
grpcio==1.16.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
opencv-python==3.4.3.18
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.7
requests==2.20.1
rsa==4.0
six==1.11.0
urllib3==1.24.1
waitress==1.1.0
Werkzeug==0.14.1